Pre Transition to meters
Includes a statically defined collar and collar component, but still an issue with port construction that allows global rotation.  Also, full collar processing needs to be implemented.  CQ2b (in meters not mm) model is added but not yet used.  package.json was also updated for drei, fiber, three, react, and react dom.